Task-3/requirements.txt
kafka-python
scikit-learn
joblib
numpy
pandas
